+++ py/desitarget/test/test_secondary.py
@@ -2,8 +2,12 @@
 # -*- coding: utf-8 -*-
 """Test desitarget secondary targets.
 """
+import os
+import numpy as np
 import unittest
 from importlib import import_module
+from glob import glob
+from pkg_resources import resource_filename
 
 
 class TestSECONDARY(unittest.TestCase):
@@ -11,26 +15,46 @@ class TestSECONDARY(unittest.TestCase):
     def setUp(self):
         # ADM these are the allowed types of observations of secondaries.
         self.flavors = {"SPARE", "DEDICATED", "SSV", "QSO", "TOO"}
+        # ADM this is the list of defined directories for SV.
+        fns = glob(resource_filename('desitarget', 'sv*'))
+        svlist = [os.path.basename(fn) for fn in fns if os.path.isdir(fn)]
+        # ADM loop over all SV flavors and main survey to get all masks.
+        from desitarget.targetmask import scnd_mask as Mx
+        self.Mxs = [Mx]
+        for sv in svlist:
+            targmask = import_module(
+                "desitarget.{}.{}_targetmask".format(sv, sv))
+            self.Mxs.append(targmask.scnd_mask)
 
     def test_flavors(self):
         """Test that secondary masks only have the allowed flavors.
         """
-        # ADM loop over flavors of SV and the main survey to get masks.
-        from desitarget.targetmask import scnd_mask as Mx
-        Mxs = [Mx]
-        svs = 2
-        for i in range(1, svs):
-            targmask = import_module(
-                "desitarget.sv{}.sv{}_targetmask".format(i, i))
-            Mxs.append(targmask.scnd_mask)
-
         # ADM for each mask...
-        for Mx in Mxs:
+        for Mx in self.Mxs:
             # ADM ...if we've already defined the flavor property...
             if "flavor" in dir(Mx[Mx[0]]):
                 flavs = set([Mx[bitname].flavor for bitname in Mx.names()])
                 # ADM ...all of the included flavors are allowed flavors.
                 self.assertTrue(flavs.issubset(self.flavors))
+
+    def test_downsample(self):
+        """Test that secondary masks all have downsample defined and <= 1.
+        """
+        # ADM for each mask...
+        for Mx in self.Mxs:
+            try:
+                # ADM create a list of all of the downsample values.
+                ds = []
+                for bitname in Mx.names():
+                    if "TOO" not in bitname:
+                        ds.append(Mx[bitname].downsample)
+            except AttributeError:
+                # ADM check downsample is defined for all scnd_mask bits.
+                msg = "downsample missing for bit {} in mask {}".format(
+                    bitname, Mx._name)
+                raise AttributeError(msg)
+        # ADM check downsample is always less than 1 (< 100%).
+        self.assertTrue(np.all(np.array(ds) <= 1))
 
 
 if __name__ == '__main__':