+++ doc/changes.rst
@@ -5,6 +5,17 @@ desitarget Change Log
 0.50.1 (unreleased)
 -------------------
 
+* Main Survey, secondary, outside-footprint target updates [`PR #677`_]:
+    * Update priorities for some secondary programs.
+    * MTL changes to reobserve 0.7 < z < 2.1 QSOs at low priority.
+        * Requested by two secondary programs.
+    * Allow secondaries to be truncated to first N targets in a file.
+    * Add fluxes/mags to the file of "standalone" secondaries.
+    * Apply bright star mask and a mag > 16 limit to secondary targets.
+    * Mask Main Survey MWS targets on BRIGHT and GALAXY using MASKBITS.
+    * Debug reading MTL ledger headers, now that they have units.
+    * Apply non-linear correction to de-extinct Gaia magnitudes TWICE.
+    * Add ``PHOT_G_N_OBS`` to the Gaia edr3 data model.
 * Add formalism for Targets of Opportunity [`PR #676`_]. Includes:
     * New ToO module.
     * Function and script to write an initial ToO .ecsv ledger.
@@ -14,6 +25,7 @@ desitarget Change Log
 
 .. _`PR #675`: https://github.com/desihub/desitarget/pull/675
 .. _`PR #676`: https://github.com/desihub/desitarget/pull/676
+.. _`PR #677`: https://github.com/desihub/desitarget/pull/677
 
 
 0.50.0 (2021-01-29)